Added a LICENSE file so that it's clear in GitHub.
GitHub exposes links to the license from the main page, so it's useful
to have this clear at the top that it is licensed under multiple
licenses. They are not compatible - the whole is not under the CDDL.
But they can be distributed within the repository without violation
of the licenses.