Changes to PGmod to build cross compiling; created filer component.
The PGmod library now builds as LanMan98PGLib. This is then used by the
filer component so that it can build an application. The application
should have all the bits that the top level !LanMan98 had. However it
seems to hang the machine right now, so I must have built something
wrong!