Fix for hang on loading the application.
The LanMan98 module was being linked with the non-module libraries.
That's so very bad - it causes the library data to be in different
places in memory and for different operations to do the wrong things.
It's unsurprising it hung (although it'd have been been nice if it
had actually crashed). Discovered by examination in Pyromaniac -
using the trace.switraps to trap the SWIs into the module showed that
the code was accessing variables differently in different parts of
the code.

The network code in BaseLib wasn't being built with -DTCP5 as it had
expected. This probably made no difference, but is at least consistent
with the original now.

Created a top level makefile which can build everything in one go.
Building with `riscos-amu` will perform an incremental build on all
the parts. The `clean` target will just clean all the libraries,
modules and application. The `fresh` target will clean and then build
the lot. `riscos-amu fresh` takes about 6 seconds on my macBook,
which makes it very useable for development.

The PGLib has been updated to include error reports if the templates
fail to open or Wimp_Initialise fails. This is useful if you have
a system that doesn't have a WindowManager present, or you forgot to
put the templates in the right place. Otherwise it is much later
in the code that it dies after writing to zero page.