Add automatic builds through GitHub actions/RISC OS Build service.
The build has been tailored very slightly to build using the RISC OS
build service (largely the changes were just in making a working
robuild.yaml file). This allows it to be built automatically on push,
although it takes a lot longer under emulation than it would on a
real RISC OS machine, or in cross-compilation.