Add .gitattributes to make source types properly.
For GitLab and GitHub there are markers to add into the .gitattributes
file to show what language (and syntax) should be used for those files.
The attributes here should allow us to colour the text properly in
both GitLab and GitHub's rendering of the source.

Update the gitignore, using the form preferred by riscos-project tool.